Move all API types and parsers from utils to adapters
