+++ ui/src/adapters/api/index.ts
@@ -1,0 +1,84 @@
+import axios from "axios";
+import z from "zod";
+
+import { StrictSchema } from "src/utils/types";
+
+export interface APIError {
+  message: string;
+  status?: number;
+}
+
+interface APIErrorResponse {
+  error: APIError;
+  isSuccessful: false;
+}
+
+interface APISuccessfulResponse<D> {
+  data: D;
+  isSuccessful: true;
+}
+
+export type APIResponse<D> = APISuccessfulResponse<D> | APIErrorResponse;
+
+export enum HTTPStatusError {
+  Aborted = 0,
+  Unauthorized = 401,
+  NotFound = 404,
+}
+
+export enum HTTPStatusSuccess {
+  Accepted = 202,
+  Created = 201,
+  NoContent = 204,
+  OK = 200,
+}
+
+export interface ResultOK<D> {
+  data: D;
+  status: HTTPStatusSuccess.OK;
+}
+
+export interface ResultCreated<D> {
+  data: D;
+  status: HTTPStatusSuccess.Created;
+}
+
+export interface ResponseError {
+  data: { message: string };
+  status: number;
+}
+
+export const responseError = StrictSchema<ResponseError>()(
+  z.object({
+    data: z.object({ message: z.string() }),
+    status: z.number(),
+  })
+);
+
+export function buildAPIError(error: unknown): APIError {
+  if (axios.isCancel(error)) {
+    return { message: error.toString(), status: HTTPStatusError.Aborted };
+  }
+
+  if (axios.isAxiosError(error)) {
+    try {
+      // This is a Polygon ID API error.
+      const { data, status } = responseError.parse(error.response);
+      const { message } = data;
+
+      return { message, status };
+    } catch (e) {
+      // This catches a CORS or other network error.
+      return { message: error.message };
+    }
+  }
+
+  if (error instanceof Error) {
+    // This is an application-level error.
+    return { message: error.toString() };
+  }
+
+  // This shouldn't happen (catch-all).
+  console.error(error);
+  return { message: "Unknown error" };
+}
