+++ ui/src/components/credentials/LinkTable.tsx
@@ -17,6 +17,7 @@ import { useCallback, useEffect, useState } from "react";
 import { useSearchParams } from "react-router-dom";
 import { z } from "zod";
 
+import { APIError } from "src/adapters/api";
 import { Credential, credentialUpdate, credentialsGetAll } from "src/adapters/api/credentials";
 import { Schema } from "src/adapters/api/schemas";
 import { ReactComponent as IconDots } from "src/assets/icons/dots-vertical.svg";
@@ -25,7 +26,7 @@ import { ErrorResult } from "src/components/shared/ErrorResult";
 import { NoResults } from "src/components/shared/NoResults";
 import { TableCard } from "src/components/shared/TableCard";
 import { useEnvContext } from "src/contexts/env";
-import { APIError, processZodError } from "src/utils/adapters";
+import { processZodError } from "src/utils/adapters";
 import { isAbortedError, makeRequestAbortable } from "src/utils/browser";
 import { ACCESSIBLE_UNTIL, LINKS, QUERY_SEARCH_PARAM } from "src/utils/constants";
 import { formatDate } from "src/utils/forms";
