+++ ui/src/components/schemas/SchemasTable.tsx
@@ -3,6 +3,7 @@ import { ColumnsType } from "antd/es/table";
 import { useCallback, useEffect, useState } from "react";
 import { Link, generatePath, useSearchParams } from "react-router-dom";
 
+import { APIError } from "src/adapters/api";
 import { Schema, getSchemas } from "src/adapters/api/schemas";
 import { ReactComponent as IconSchema } from "src/assets/icons/file-search-02.svg";
 import { ReactComponent as IconUpload } from "src/assets/icons/upload-01.svg";
@@ -11,7 +12,7 @@ import { NoResults } from "src/components/shared/NoResults";
 import { TableCard } from "src/components/shared/TableCard";
 import { useEnvContext } from "src/contexts/env";
 import { ROUTES } from "src/routes";
-import { APIError, processZodError } from "src/utils/adapters";
+import { processZodError } from "src/utils/adapters";
 import { isAbortedError, makeRequestAbortable } from "src/utils/browser";
 import { IMPORT_SCHEMA, QUERY_SEARCH_PARAM, SCHEMAS, SCHEMA_TYPE } from "src/utils/constants";
 import { formatDate } from "src/utils/forms";
